requirements.txt
asgiref==3.7.2
Brotli==1.1.0
cffi==1.16.0
click==8.1.7
crispy-bootstrap5==2023.10
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.2
django-crispy-forms==2.1
django-import-export==3.3.7
django-weasyprint==2.3.0
et-xmlfile==1.1.0
fonttools==4.49.0
h11==0.14.0
html5lib==1.1
MarkupPy==1.14
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
pillow==10.2.0
plotly==5.19.0
pycparser==2.21
pydyf==0.8.0
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tenacity==8.2.3
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.27.1
weasyprint==61.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.3
